+++ IDisposableAnalyzers.Test/IDISP011DontReturnDisposedTests/Valid.cs
@@ -1167,7 +1167,7 @@ namespace N
             }
 
             process.Start();
-            return process;  // IDISP011	Don't return disposed instance.
+            return process;  // IDISP011	Don't return disposed instance
         }
     }
 }